eliminating the use of MinDFA, replacing with regex matches check  Signed-off-by: shireenf-ibm <shireenf@il.ibm.com>
